+++ wasm/Cargo.toml
@@ -1,15 +1,13 @@
 [package]
 name = "parside-wasm"
+description = "Parser library to handle Composable Event Streaming Representation (CESR)"
+license = "Apache-2.0"
+keywords = ["cesr", "keri", "acdc"]
 version = "0.1.0"
-authors = ['Dmitry Kuzmenko <dmitry.kuzmenko@dsr-corporation.com>']
 edition = "2021"
 
 [lib]
-crate-type = ["cdylib", "rlib"]
-
-[dependencies.getrandom]
-version = '0.2'
-features = ['js']
+crate-type = ["cdylib"]
 
 [dependencies.parside_core]
 path = '..'
@@ -17,21 +15,17 @@ package = "parside"
 
 [dependencies.cesride_wasm]
 git = "https://github.com/DSRCorporation/cesride.git"
-branch = "wasm-parside"
+branch = "fix/wasm-readme"
 package = "cesride-wasm"
 
 [dependencies.cesride]
 git = "https://github.com/DSRCorporation/cesride.git"
-branch = "wasm-parside"
-
-[dependencies.wasm-bindgen]
-version = '0.2'
-features = ['serde-serialize']
-
-[dependencies.web-sys]
-version = "0.3"
-features = ["console", ]
+branch = "fix/wasm-readme"
 
 [dependencies]
+getrandom = { version = "0.2", features = ['js'] }
 js-sys = '0.3'
-serde_json = "1.0"
+serde = { version = "1.0", features = ["derive"] }
+serde-wasm-bindgen = "0.4"
+wasm-bindgen = { version = "0.2", features = ['serde-serialize'] }
+web-sys = { version = "0.3", features = ["console", ] }
