+++ wasm/demo/node/package.json
@@ -1,7 +1,7 @@
 {
   "name": "parside-wasm-demo",
   "version": "0.1.0",
-  "description": "JS demo for parside wasm bindings",
+  "description": "NodeJS demo for parside wasm bindings",
   "scripts": {
     "start": "ts-node src/main.ts",
     "check": "tslint -c tslint.json 'src/**/*.ts'"
