--- wasm/package.json
@@ -1,14 +1,0 @@
-{
-  "scripts": {
-    "build": "webpack",
-    "serve" : "webpack serve"
-  },
-  "devDependencies": {
-    "@wasm-tool/wasm-pack-plugin": "1.5.0",
-    "text-encoding": "^0.7.0",
-    "html-webpack-plugin": "^5.3.2",
-    "webpack": "^5.49.0",
-    "webpack-cli": "^4.7.2",
-    "webpack-dev-server": "^3.11.2"
-  }
-}
