--- wasm/rustfmt.toml
@@ -1,2 +1,0 @@
-reorder_imports = true
-reorder_modules = true
