+++ wasm/src/message/groups/controller_idx_sigs.rs
@@ -77,6 +77,7 @@ impl ControllerIdxSigWrapper {
         ControllerIdxSigWrapper(ControllerIdxSig::new((*siger).clone()))
     }
 
+    #[wasm_bindgen(getter)]
     pub fn siger(&self) -> SigerWrapper {
         SigerWrapper::wrap(&self.0.siger)
     }
