+++ wasm/src/message/groups/pathed_material_quadlets.rs
@@ -67,6 +67,7 @@ impl PathedMaterialQuadletWrapper {
         PathedMaterialQuadletWrapper(PathedMaterialQuadlet::new((*siger).clone()))
     }
 
+    #[wasm_bindgen(getter)]
     pub fn siger(&self) -> SigerWrapper {
         SigerWrapper::wrap(&self.0.siger)
     }
