+++ wasm/src/message/groups/sad_path_sig.rs
@@ -65,6 +65,7 @@ impl SadPathSigWrapper {
         SadPathSigWrapper(SadPathSig::new((*siger).clone()))
     }
 
+    #[wasm_bindgen(getter)]
     pub fn siger(&self) -> SigerWrapper {
         SigerWrapper::wrap(&self.0.siger)
     }
