+++ wasm/src/message/groups/trans_receipt_quadruples.rs
@@ -77,15 +77,22 @@ impl TransReceiptQuadrupleWrapper {
         ))
     }
 
+    #[wasm_bindgen(getter)]
     pub fn prefixer(&self) -> PrefixerWrapper {
         PrefixerWrapper::wrap(&self.0.prefixer)
     }
+
+    #[wasm_bindgen(getter)]
     pub fn seqner(&self) -> SeqnerWrapper {
         SeqnerWrapper::wrap(&self.0.seqner)
     }
+
+    #[wasm_bindgen(getter)]
     pub fn saider(&self) -> SaiderWrapper {
         SaiderWrapper::wrap(&self.0.saider)
     }
+
+    #[wasm_bindgen(getter)]
     pub fn siger(&self) -> SigerWrapper {
         SigerWrapper::wrap(&self.0.siger)
     }
