+++ wasm/src/message/mod.rs
@@ -1,5 +1,4 @@
-pub mod cold_code;
-// pub mod custom_payload;
+pub mod custom_payload;
 pub mod groups;
 pub mod message;
 pub mod message_list;
